BackEnd/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
cachetools==5.5.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==4.2.4
django-allauth==0.51.0
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
environ==1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
huggingface-hub==0.26.2
idna==3.4
inflection==0.5.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.47
propcache==0.2.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.13
yarl==1.17.2
